backend/requirements.txt
# SHADOW-7 Backend Dependencies
# Starlette 0.40+ required for max_part_size (large file uploads)
fastapi>=0.115.0,<0.120.0
starlette>=0.40.0
uvicorn[standard]==0.27.0
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
httpx==0.26.0
python-dotenv==1.0.0
aiofiles==23.2.1
mammoth==1.6.0
weasyprint==60.2
reportlab==4.0.9
Pillow==10.2.0
python-docx==1.1.0
ebooklib==0.18
google-generativeai>=0.8.0
